pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "arc-finetuning-st"
version = "0.1.0"
description = ""
authors = ["Andrei Fajardo <andrei@llamaindex.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
llama-index-core = "^0.11.11"
llama-index-finetuning = "^0.2.1"
llama-index-llms-openai = "^0.2.7"
llama-index-program-openai = "^0.2.0"
streamlit = "^1.38.0"
plotly = "^5.24.1"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
instructor = "^1.5.2"
litellm = {extras = ["proxy"], version = "^1.49.2"}

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.2.5"
llama-index-utils-workflow = "^0.2.1"

[tool.poetry.scripts]
arc-finetuning-cli = 'arc_finetuning_st.cli.command_line:main'
